requirements.txt
AWSIoTPythonSDK
matplotlib
python-dotenv
numpy
pyrtlsdr
